requirements.txt
pytest
tfx==1.9.1
kfp
tensorflow-cloud
keras-tuner
PyGithub
wget